requirements.txt
astroid==2.0.4
asttokens==1.1.12
birdseye==0.7.2
cached-property==1.5.1
cheap-repr==0.2.0
Click==7.0
dataclasses-json==0.0.22
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
Flask==1.0.2
Flask-Humanize==0.3.0
future==0.17.1
humanize==0.5.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
littleutils==0.2.1
Markdown==3.0.1
MarkupSafe==1.0
marshmallow==3.0.0b20
mccabe==0.6.1
outdated==0.1.2
pylint==2.1.1
pytz==2018.7
qualname==0.1.0
six==1.11.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
wrapt==1.10.11
